Decompile `sat` `InitStatsAndGear`
