+++ config/saturn/game_user_syms.txt
@@ -13,6 +13,7 @@ ___ashlsi3 = 0x06030F0C;
 _rand = 0x06032E68;
 _DAT_06057f62 = 0x06057F62;
 _DAT_0605c10c = 0x0605C10C;
+_D_8003C730 = 0x0605C11C;
 _g_Player = 0x0605C140;
 _currentMusicId = 0x0605C5AC;
 _g_RoomCount = 0x0605C5B0;
@@ -24,8 +25,10 @@ _g_CastleFlags = 0x0605CA50;
 _g_GameTimer = 0x0605CD50;
 _D_8006BB00 = 0x0605CEB4;
 _D_8013AEE4 = 0x0605CEB8;
+_g_CastleMap = 0x0605CEC0;
 _g_randomNext = 0x0605D748;
 _g_StageId = 0x0605D750;
+_g_GameClearFlag = 0x0605D784;
 _g_Servant = 0x0605D7E4;
 _D_8003C708 = 0x0605D7E8;
 _g_MenuNavigation = 0x0605D790;
@@ -38,6 +41,7 @@ _CalcDefense = 0x0606DCF0;
 _CalcPlayerDamage = 0x0606F2C0;
 _CheckEquipmentItemCount = 0x0606F448;
 _AddToInventory = 0x0606F4C4;
+_UpdateCapePalette = 0x060749F8;
 _MoveEntity = 0x06079B9C;
 _GetDistanceToPlayerX = 0x0607A9F8;
 _GetDistanceToPlayerY = 0x0607AA1C;
@@ -49,6 +53,7 @@ _SetSubStep = 0x0607B240;
 _CreateEntityFromCurrentEntity = 0x0607B318;
 _AnimateEntity = 0x0607B674;
 _CheckCollision = 0x0607B7B4;
+_MTH_GetRand = 0x0607C0BC;
 _g_SubwpnDefs = 0x0607C14C;
 _g_EquipDefs = 0x0607C250;
 _g_AccessoryDefs = 0x0607F0A0;
@@ -61,6 +66,9 @@ _g_ButtonMask = 0x060854C8;
 _g_LevelHPIncrease = 0x060854E0;
 _D_801375C8 = 0x06085524;
 _DAT_06085c70 = 0x06085C70;
+_g_LuckCode = 0x06085C90;
+_g_AxeArmorCode = 0x06085C94;
+_g_GTIClubCode = 0x06085C98;
 _g_SineTable = 0x06085E86;
 _DAT_0608609c = 0x0608609C;
 _D_80137960 = 0x060860A0;
@@ -70,5 +78,8 @@ _DAT_060860D8 = 0x060860D8;
 _g_StatBuffTimers = 0x060861B0;
 _D_801375CC = 0x06086238;
 _D_8013B5E8 = 0x06086380;
+_D_801397FC = 0x06086384;
+_D_80139008 = 0x0608638C;
+_DAT_06086390 = 0x06086390;
 _g_Entities = 0x060997F8;
 _DAT_06099811 = 0x06099811;
