+++ src/saturn/sattypes.h
@@ -271,14 +271,15 @@ typedef struct {
 
 // offsets are not the same as psx
 typedef struct {
-    u8 relics[30];
-    char pad4[0xc];
+    u8 relics[32];
+    char pad[2];
+    u8 spells[8];
     u8 equipHandCount[176];
-    u8 equipBodyCount[26];
-    char pad5[0x42];
+    u8 equipBodyCount[92];
     u8 equipHandOrder[176];
-    u8 equipBodyOrder[26];
-    char pad2[0x4E];
+    u8 equipBodyOrder[92];
+    u8 saveName[8];
+    u32 spellsLearnt;
     s32 hp;
     s32 hpMax;
     s32 hearts;
@@ -290,7 +291,9 @@ typedef struct {
     s32 statsTotal[4];
     s32 level;
     u32 exp;
-    char pad3[0xC];
+    s32 gold;
+    s32 killCount;
+    u32 D_80097BF8;
     u32 subWeapon;
     u32 equipment[8];
     u32 attackHands[2];
@@ -497,7 +500,7 @@ typedef struct {
 } unkGraphicsStruct;
 
 typedef struct {
-    s32 : 32;
+    const char* name;
     s32 : 32;
     u16 : 16;
     u16 : 16;
