+++ tools/builds/saturn.py
@@ -92,6 +92,7 @@ snd_srcs = [
     'src/saturn/zero.c',
     'src/saturn/lib/snd.c',
     'src/saturn/zero_2.c',
+    'src/saturn/game_0.c',
     'src/saturn/game.c',
     'src/saturn/richter.c',
     'src/saturn/stage_02.c',
@@ -155,7 +156,6 @@ def link_objs(srcs, output_dir):
                 'syms_file': syms_file})
 
 objs = [
-    'build/saturn/game.o',
     'build/saturn/alucard.o',
     'build/saturn/richter.o',
     'build/saturn/stage_02.o',
@@ -182,6 +182,9 @@ def link_multi(multi_objs, output_dir):
                 'objs': sub_objs})
 
 multi_objs = {
+    'build/saturn/game.o' : [
+        'build/saturn/game_0.o',
+    ],
     'build/saturn/zero.o' : [
         'build/saturn/lib/snd.o',
         'build/saturn/zero_2.o',
